// frontend/src/components/sidebar/types.ts
export interface ContextMenuItem {
  type?: 'separator'
  label?: string
  onClick?: () => void
  danger?: boolean
  disabled?: boolean
  shortcut?: string
  children?: ContextMenuItem[]
}